# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrainChat API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  createdAt: string;
  onboardingComplete: boolean;
}

export interface AuthResponse {
  user: User;
}

export interface CreateProfileBody {
  trainingGoal: string;
  experienceLevel: string;
  trainingStyle: string;
  daysPerWeek: number;
  sessionDuration: number;
  equipmentAccess: string;
  /** @nullable */
  injuries?: string | null;
  /** @nullable */
  sportFocus?: string | null;
  /** @nullable */
  exercisePreferences?: string | null;
  /** @nullable */
  exercisesToAvoid?: string | null;
}

export interface UserProfile {
  id: number;
  userId: number;
  trainingGoal: string;
  experienceLevel: string;
  trainingStyle: string;
  daysPerWeek: number;
  sessionDuration: number;
  equipmentAccess: string;
  /** @nullable */
  injuries?: string | null;
  /** @nullable */
  sportFocus?: string | null;
  /** @nullable */
  exercisePreferences?: string | null;
  /** @nullable */
  exercisesToAvoid?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateConversationBody {
  title: string;
}

export interface Conversation {
  id: number;
  userId: number;
  title: string;
  createdAt: string;
  updatedAt: string;
  messageCount: number;
  /** @nullable */
  lastMessage?: string | null;
}

export interface SendMessageBody {
  content: string;
}

export interface Message {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
  /** @nullable */
  structuredData?: string | null;
}

export interface MessagePair {
  userMessage: Message;
  assistantMessage: Message;
}

export interface CreateExerciseBody {
  name: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: string | null;
  /** @nullable */
  rest?: string | null;
  /** @nullable */
  notes?: string | null;
  orderIndex: number;
}

export interface CreateProgramDayBody {
  dayNumber: number;
  name: string;
  exercises: CreateExerciseBody[];
  /** @nullable */
  notes?: string | null;
}

export interface CreateProgramBody {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  conversationId?: number | null;
  days: CreateProgramDayBody[];
}

export interface Program {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  conversationId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface Exercise {
  id: number;
  programDayId: number;
  name: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: string | null;
  /** @nullable */
  rest?: string | null;
  /** @nullable */
  notes?: string | null;
  orderIndex: number;
}

export interface ProgramDay {
  id: number;
  programId: number;
  dayNumber: number;
  name: string;
  /** @nullable */
  notes?: string | null;
  exercises: Exercise[];
}

export interface ProgramDetail {
  id: number;
  userId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  conversationId?: number | null;
  createdAt: string;
  updatedAt: string;
  days: ProgramDay[];
}