# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrainChat API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: profile
    description: User profile operations
  - name: conversations
    description: Conversation operations
  - name: messages
    description: Message operations
  - name: programs
    description: Training program operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /profile:
    get:
      operationId: getProfile
      tags: [profile]
      summary: Get user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProfile
      tags: [profile]
      summary: Create or update user profile (onboarding)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfileBody"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /conversations:
    get:
      operationId: listConversations
      tags: [conversations]
      summary: List user conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
    post:
      operationId: createConversation
      tags: [conversations]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationBody"
      responses:
        "201":
          description: Conversation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"

  /conversations/{id}:
    get:
      operationId: getConversation
      tags: [conversations]
      summary: Get a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteConversation
      tags: [conversations]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /conversations/{id}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message and get AI response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "200":
          description: AI response message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagePair"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List saved training programs
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Save a training program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramBody"
      responses:
        "201":
          description: Program created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"

  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get a training program with days
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program with days
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProgramDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProgram
      tags: [programs]
      summary: Delete a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required:
        - email
        - password
        - name

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        createdAt:
          type: string
        onboardingComplete:
          type: boolean
      required:
        - id
        - email
        - name
        - createdAt
        - onboardingComplete

    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
      required:
        - user

    CreateProfileBody:
      type: object
      properties:
        trainingGoal:
          type: string
        experienceLevel:
          type: string
        trainingStyle:
          type: string
        daysPerWeek:
          type: integer
        sessionDuration:
          type: integer
        equipmentAccess:
          type: string
        injuries:
          type: ["string", "null"]
        sportFocus:
          type: ["string", "null"]
        exercisePreferences:
          type: ["string", "null"]
        exercisesToAvoid:
          type: ["string", "null"]
      required:
        - trainingGoal
        - experienceLevel
        - trainingStyle
        - daysPerWeek
        - sessionDuration
        - equipmentAccess

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        trainingGoal:
          type: string
        experienceLevel:
          type: string
        trainingStyle:
          type: string
        daysPerWeek:
          type: integer
        sessionDuration:
          type: integer
        equipmentAccess:
          type: string
        injuries:
          type: ["string", "null"]
        sportFocus:
          type: ["string", "null"]
        exercisePreferences:
          type: ["string", "null"]
        exercisesToAvoid:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - trainingGoal
        - experienceLevel
        - trainingStyle
        - daysPerWeek
        - sessionDuration
        - equipmentAccess
        - createdAt
        - updatedAt

    CreateConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title

    Conversation:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
        messageCount:
          type: integer
        lastMessage:
          type: ["string", "null"]
      required:
        - id
        - userId
        - title
        - createdAt
        - updatedAt
        - messageCount

    SendMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content

    Message:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
        structuredData:
          type: ["string", "null"]
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt

    MessagePair:
      type: object
      properties:
        userMessage:
          $ref: "#/components/schemas/Message"
        assistantMessage:
          $ref: "#/components/schemas/Message"
      required:
        - userMessage
        - assistantMessage

    CreateProgramBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        conversationId:
          type: ["integer", "null"]
        days:
          type: array
          items:
            $ref: "#/components/schemas/CreateProgramDayBody"
      required:
        - name
        - days

    CreateProgramDayBody:
      type: object
      properties:
        dayNumber:
          type: integer
        name:
          type: string
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/CreateExerciseBody"
        notes:
          type: ["string", "null"]
      required:
        - dayNumber
        - name
        - exercises

    CreateExerciseBody:
      type: object
      properties:
        name:
          type: string
        sets:
          type: ["integer", "null"]
        reps:
          type: ["string", "null"]
        rest:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        orderIndex:
          type: integer
      required:
        - name
        - orderIndex

    Program:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        conversationId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - userId
        - name
        - createdAt
        - updatedAt

    ProgramDay:
      type: object
      properties:
        id:
          type: integer
        programId:
          type: integer
        dayNumber:
          type: integer
        name:
          type: string
        notes:
          type: ["string", "null"]
        exercises:
          type: array
          items:
            $ref: "#/components/schemas/Exercise"
      required:
        - id
        - programId
        - dayNumber
        - name
        - exercises

    Exercise:
      type: object
      properties:
        id:
          type: integer
        programDayId:
          type: integer
        name:
          type: string
        sets:
          type: ["integer", "null"]
        reps:
          type: ["string", "null"]
        rest:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        orderIndex:
          type: integer
      required:
        - id
        - programDayId
        - name
        - orderIndex

    ProgramDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        conversationId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        days:
          type: array
          items:
            $ref: "#/components/schemas/ProgramDay"
      required:
        - id
        - userId
        - name
        - createdAt
        - updatedAt
        - days