# lib/api-zod/src/generated/types/createExerciseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrainChat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateExerciseBody {
  name: string;
  /** @nullable */
  sets?: number | null;
  /** @nullable */
  reps?: string | null;
  /** @nullable */
  rest?: string | null;
  /** @nullable */
  notes?: string | null;
  orderIndex: number;
}