# lib/api-zod/src/generated/types/createProgramBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrainChat API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateProgramDayBody } from "./createProgramDayBody";

export interface CreateProgramBody {
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  conversationId?: number | null;
  days: CreateProgramDayBody[];
}