# lib/api-zod/src/generated/types/createProgramDayBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrainChat API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateExerciseBody } from "./createExerciseBody";

export interface CreateProgramDayBody {
  dayNumber: number;
  name: string;
  exercises: CreateExerciseBody[];
  /** @nullable */
  notes?: string | null;
}