# lib/api-zod/src/generated/types/programDay.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrainChat API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Exercise } from "./exercise";

export interface ProgramDay {
  id: number;
  programId: number;
  dayNumber: number;
  name: string;
  /** @nullable */
  notes?: string | null;
  exercises: Exercise[];
}