# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrainChat API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  userId: number;
  trainingGoal: string;
  experienceLevel: string;
  trainingStyle: string;
  daysPerWeek: number;
  sessionDuration: number;
  equipmentAccess: string;
  /** @nullable */
  injuries?: string | null;
  /** @nullable */
  sportFocus?: string | null;
  /** @nullable */
  exercisePreferences?: string | null;
  /** @nullable */
  exercisesToAvoid?: string | null;
  createdAt: string;
  updatedAt: string;
}